Initial import of saxon-6.2.2, another java XSLT implementation.
from pkg/DESCR:

The SAXON package is a collection of tools for processing XML
documents. The main components are:

* An XSLT processor, which implements the Version 1.0 XSLT and
XPath Recommendations from the World Wide Web Consortium.  This
version of Saxon also includes some features defined in XSLT 1.1.

* A Java library, which supports a similar processing model to XSL,
but allows full programming capability, which you need if you want
to perform complex processing of the data or to access external
services such as a relational database

* A slightly improved version of the AElfred parser from Microstar.
(But you can use SAXON with any SAX-compliant XML parser if you
prefer).

Saxon is distributed under the terms of the Mozilla Public License
(MPL).