courier-imap: IMAP server for access to Maildir-style mailboxes
Courier-IMAP is a server that provides IMAP access to Maildirs. This IMAP
server does NOT handle traditional mailbox files (/var/spool/mail, and
derivatives), it was written for the specific purpose of providing IMAP
access to Maildirs.

"Maildir" is a directory-based mail storage format originally introduced
in the Qmail mail server, and adopted as an alternative mail storage
format by both Exim and Postfix. Courier-IMAP is popular on
Qmail/Exim/Postfix sites that are configured to use maildirs. The primary
advantage of maildirs is that multiple applications can access the same
Maildir simultaneously without requiring any kind of locking whatsoever.
It is a faster and more efficient way to store mail. It works particularly
well over NFS, which has a long history of locking-related woes.