initial import of GnuCap
GnuCap is a general purpose circuit simulator.  GnuCap was
formerly known as ACS.  GnuCap performs nonlinear
dc and transient analyses, fourier analysis, and ac analysis
linearized at an operating point.  It is fully interactive and
command driven.  It can also be run in batch mode or as a server.
The output is produced as it simulates.  Spice compatible models
for the MOSFET (level 1-7) and diode are included in this
release.
 
Since it is fully interactive, it is possible to make changes and
re-simulate quickly.  The interactive design makes it well suited
to the typical iterative design process used it optimizing a circuit
design.

Unlike Spice, the engine is designed to do true mixed-mode
simulation. Most of the code is in place for future support of
event driven analog simulation, and true multi-rate simulation.
	  
If you are tired of Spice and want a second opinion, you want to
play with the circuit and want a simulator that is interactive,
you want to study the source code and want something easier to
follow than Spice, or you are a researcher working on modeling
and want automated model generation tools to make your job easier,
try GnuCap.