Initial import of flan-1.1beta, as flan-1.1rc1, into the NetBSD Packages Collection.
Provided in PR 15687 by Ola Eriksson (ola@mrEriksson.net), modified by myself.

Flan is a network simluator written in java, and is now fully
functionnal.  The IP protocol is implemented, with a quite complete
RIPv2 routing algorithm.  The protocols are also written in java, so
GO AHEAD !  Flan is under GNU Public License, so download it, modify
it, and distribute it. 

The networks are now encoded in XML, using XMLCoDe.  This very useful
library (included in the flan tarball as jars) automatically creates
an XML image of a java object, like the flan nets (XMLCoDe is entirely
written by Sylvain Guerin).