Add st: non-preemptive thread library for Internet applications
The State Threads Library is a small application library which
provides a foundation for writing fast and highly scalable Internet
applications (such as web servers, proxy servers, mail transfer
agents, and so on, really any network-data-driven application) on
UNIX-like platforms.  It combines the simplicity of the multithreaded
programming paradigm, in which one thread supports each simultaneous
connection, with the performance and scalability of an event-driven
state machine architecture.  In other words, this library offers a
threading API for structuring an Internet application as a state
machine.