import pcb-current-20030815.
Pcb - X11 interactive printed circuit board layout system.

Pcb is a handy tool for	the X Window System to design printed
circuit	boards.  All coordinate units are 1/1000 inch.

Please note that this package is a development snapshot and while it contains
the latest and greatest features, it may be buggy as well.  There is a seperate
pcb package which is made of the stable releases.