There are three different parts we can add to the common transforming cache to speed things up:
	passthru	automatically accept certain options without
			further checking

	transform	change -[IL]{${X11BASE},${LOCALBASE},${DEPOTBASE}/*}
			to the correct ${BUILDLINK_DIR} reference

	block		block certain options without further checking

Seeding the cache appropriately lets us handle large classes of options
without having to go through the wrapper logic file, which generates
a cache hit for every single option and bloats the cache quite bit more.
The implementation is at the bottom of bsd.buildlink3.mk where
${_BLNK_WRAP_CACHE_ADD_TRANSFORM} is generated.

For now, only do the passthru stuff, since I'm not sure of the correctness
of the other two phases.  "transform" is probably okay, but "block" may be
a tad too aggressive.