Initial import of libtcpa-1.1b, a library and example programs for the TCPA chip described in IBM Global Security Analysis Lab's article "Take Control of TCPA" in the August 2003 issue of Linux Journal.
For this package to be useful, you need a computer with a TCPA
chip, and support for the chip in your kernel. An unofficial NetBSD
TCPA driver and instructions can be found here:

http://www.citi.umich.edu/u/rwash/projects/trusted/netbsd.html

I don't have a TCPA chip with which to verify the functionality of
this package.

Thanks to Soren Jacobsen for bringing me up to speed on modern
pkgsrc conventions, and to Rick Wash for his recent presentation
at my local ACM chapter on TCPA and "Trusted Computing".