Pullup "specific bulk package" functionality to the pkgsrc-2004Q1 branch.  Also pullup a minor wording change in an informational message whilst I'm here.
Requested by dmcmahill in ticket pkgsrc-14.

"Add support in the bulk build code to properly deal with SPECIFIC_PKGS=1.
In particular, when SPECIFIC_PKGS is set in /etc/mk.conf, you can now do

  sh mk/bulk/build

and have the right thing happen.  Only those packages explicitly listed
and those which are depended upon are considered for the build.  Other
than the restricted list of packages, the bulk build works the same way
as a full bulk build."