Initial import of p5-Mail-Mbox-MessageParser-0.14:
Mail::Mbox::MessageParser is a feature-poor but very fast mbox
parser. It uses the best of three strategies for parsing a mailbox:
either using cached folder information, GNU grep, or highly optimized
Perl.

XXX: might need some adaptation for OSs where diff and bzip2 are
not in the base system.