Initial import of et-2.56 from pkgsrc-wip, contributed by <dhowlland at users.sourceforge.net>.
Wolfenstein Enemy Territory is a free, stand-alone, downloadable multiplayer
game in which players wage war as Axis or Allies in team-based combat.

Enemy Territory is completely free and does not require the original
Return to Castle Wolfenstein game to play.

Enemy Territory is based on the Quake3 engine.

This package includes both the game client and the dedicated server.