Initial import of kftpgrabber 0.5.0, a graphical FTP client for KDE
There are many FTP clients around these days. Most are console based, some
feature a graphical user interface. What most clients are missing is support
for latest or not so common technologies like ZeroConf, TLS/SSL or FXP (just
to name a few), accompanied by a decent user interface, that tries to support
the user instead of frustrating him.

KFTPGrabber tries to make a complete FTP client solution with support for
plugins, scripting and everything else the users needs, based on the
latest technologies KDE provides.

Here are some of the features:

  - Multiple FTP sessions (tabs)
  - Transfer queue
  - TLS/SSL support for encrypted connections (implicit and explicit)
  - Partial X509 certificate support for authentication
  - FXP transfer support (site-to-site)
  - OTP (one time password) support - s/key, md5, rmd160, sha1
  - Drag&Drop support
  - Site bookmarking
  - Encrypted bookmark support (password can be saved to KWallet)
  - Distributed FTP support (PRET)
  - SSCN and CPSV support
  - Skiplist
  - ZeroConf (aka. Rendezvous) support for local site discovery
  - Bookmark sharing with Kopete contacts (KDE >= 3.3)
  - Bookmark import plugins
  - Support for SFTP protocol [experimental]