Import lang/postgresql80-plperl-8.0.3.
PL/Perl allows you to write functions in the Perl programming language
that may be used in SQL queries as if they were built into Postgres.
The PL/Perl intepreter is a full Perl interpreter.  However, certain
operations have been disabled in order to maintain the security of the
system.