Initial import of CSP, version 0.32, into the NetBSD Packages Collection.  If you think you might have read the openssl man pages one time too much for your own sanity, you might like this package.
Certificate Service Provider is a perl wrapper around openssl that allows you
to run multiple simple certificate authorities (CAs).  CSP is designed to be
simple (almost to a fault) and is ideally suited to small PKIs (< 1000
entities) where security is paramount.  CSP is meant to be run on isolated,
offline computers while still allowing CRLs and certificate repositories to be
easily published.

The package includes a patch that lets the program run out-of-the-box,
without setting up CSPHOME and OPENSSL in the environment.  Defining them
is of course still permitted.