Initial import of fast-user-switch-applet, version 2.14.0:
The Fast User-Switching Applet is an applet for the GNOME panel which
provides a menu to switch between users.  It integrates with GDM to
switch between existing X11 sessions or create new ones as needed,
and will show the same users as the GDM face browser.

The applet can display itself in the panel using either the current
user's name, the word "Users" or an icon, and can be configured to
include an item which launches an X session in a window.  The applet
code is designed to allow for easy cut-and-paste of the system-interaction
code into other applications, like screensavers.