OCamlDuce is an enhancement to OCaml that allows one to use XML features, such as types and values, natively in OCaml programs.
The goal of the OCamlDuce project is to extend the OCaml language with features
to make it easier to write safe and efficient complex applications that need to
deal with XML documents. In particular, it relies on a notion of types and
patterns to guarantee statically that all the possible input documents are
correctly processed, and that only valid output documents are produced.

WWW: http://www.cduce.org/ocaml