Initial import of p5-RPC-XML 0.59.
The package was originally created by Roman Kulik in pkgsrc-wip. Very
small changes by me.


The RPC::XML package is an implementation of XML-RPC. The module
provides classes for sample client and server implementations, a
server designed as an Apache location-handler, and a suite of
data-manipulation classes that are used by them.