Initial import of subtitleeditor-0.13.6. Originally packaged for pkgsrc-wip by Cesar Catrian C., updated by myself.
Subtitle Editor is a GTK+2 tool to edit subtitles. It can be used
for new subtitles or as a tool to transform, edit, correct and
refine existing subtitle. This program also shows sound waves,
which makes it easier to synchronise subtitles to voices.

Subtitle Editor is free software released under the GNU General
Public License (GPL).