Link in to native X.Org the following packages:   libFS and libXScrnSaver
For libXScrnSaver, there is a catch.  Previously, the code assumed its
presence from the fact that X11_TYPE was native.  I'm changing that to
provide a builtin.mk file.  This essentially means that the "native X.Org"
functionality of pkgsrc requires said native X.Org to provide the
necessary pkg-config files.  Expect more change of that kind.