Import tstools version 1.11.
tstools is a set of cross-platform command line tools for working with MPEG
data.

The emphasis is on relatively simple tools which concentrate on MPEG (H.264 and
H.262) data packaged according to H.222 (i.e., TS or PS), with a particular
interest in checking for conformance.

Transport Stream (TS) is typically used for distribution of cable and satellite
data. Program Stream (PS) is typically used to store data on DVDs.

The tools are focused on:

    * Quick reporting of useful data (tsinfo, stream_type)
    * Giving a quick overview of the entities in the stream (esdots, psdots)
    * Reporting on TS packets (tsreport) or ES units/frames/fields (esreport)
    * Simple manipulation of stream data (es2ts, esfilter, esreverse, esmerge,
      ts2es)
    * Streaming of data, possibly with introduced errors (tsplay)