Pullup ticket #2707 - requested by martti mediawiki: bug fix update
Revisions pulled up:
- www/mediawiki/Makefile			1.5
- www/mediawiki/distinfo			1.4
---
Module Name:	pkgsrc
Committed By:	martti
Date:		Sun Feb 22 11:58:57 UTC 2009

Modified Files:
	pkgsrc/www/mediawiki: Makefile distinfo

Log Message:
Updated www/mediawiki to 1.13.5

This is a maintenance release which corrects some bugs in the installer,
introduced during the hasty security release of 1.13.4. It is not
necessary to upgrade if you do not intend on using the installer.