Added lang/openjdk7-bin version 20090129
This is a package for the NetBSD (currently i386 5.0) binaries of
the OpenJDK 7 kit.

Many thanks to christos@ for the port and jmcneill@ for compiling up
the binaries in the first place.

This is not intended to preclude a real package that builds openjdk7
from source, but can hopefully serve as a useful bootstrap jdk for
that and other purposes.