Importing devel/p5-Perl-MinimumVersion version 1.20
Perl::MinimumVersion takes Perl source code and calculates the minimum
version of perl required to be able to run it. Because it is based on
PPI, it can do this without having to actually load the code.

Currently it tests both the syntax of your code, and the use of explicit
version dependencies such as require 5.005.