Import tex-babelbib-1.29 as print/tex-babelbib.
This package enables to generate multilingual bibliographies in
cooperation with babel.  Two approaches are possible: each citation
may be written in another language, or the whole bibliography can be
typeset in a language chosen by the user.  In addition, the package
supports commands to change the typography of the bibliographies.