Import Asterisk 1.8.1:
Asterisk is a complete PBX in software.  It provides all of the
features you would expect from a PBX and more. Asterisk does voice
over IP in three protocols, and can interoperate with almost all
standards-based telephony equipment using relatively inexpensive
hardware.

Asterisk 1.8 is a long term support version (i.e. it will be
supported for four years with an additional year of security only
fixes).  See:

https://wiki.asterisk.org/wiki/display/AST/Asterisk+Versions

     What's new:

Asterisk 1.8 is the next major release series of Asterisk.

The release of Asterisk 1.8.0 would not have been possible without the support
and contributions of the community. Since Asterisk 1.6.2, we've had over 500
reporters, more than 300 testers and greater than 200 developers contributed to
this release.

You can find a summary of the work involved with the 1.8.0 release in the
sumary:

http://svn.asterisk.org/svn/asterisk/tags/1.8.0/asterisk-1.8.0-summary.txt

A short list of available features includes:

     * Secure RTP
     * IPv6 Support in the SIP channel driver
     * Connected Party Identification Support
     * Calendaring Integration
     * A new call logging system, Channel Event Logging (CEL)
     * Distributed Device State using Jabber/XMPP PubSub
     * Call Completion Supplementary Services support
     * Advice of Charge support
     * Much, much more!

A full list of new features can be found in the CHANGES file.

http://svn.digium.com/view/asterisk/branches/1.8/CHANGES?view=markup

For a full list of changes in the current release candidate, please see the
ChangeLog:

http://downloads.asterisk.org/pub/telephony/asterisk/ChangeLog-1.8.0

-----

The Asterisk Development Team has announced the release of Asterisk 1.8.1.

The release of Asterisk 1.8.1 resolves several issues reported by the
community and would have not been possible without your participation.
Thank you!

The following is a sample of the issues resolved in this release:

* Fix issue when using directmedia. Asterisk needs to limit the codecs offered
   to just the ones that both sides recognize, otherwise they may end up sending
   audio that the other side doesn't understand.
   (Closes issue #17403. Reported, patched by one47. Tested by one47, falves11)

* Resolve issue where Party A in an analog 3-way call would continue to hear
   ringback after party C answers.
   (Patched by rmudgett)

* Fix playback failure when using IAX with the timerfd module.
   (Closes issue #18110. Reported, tested by tpanton. Patched by jpeeler)

* Fix problem with qualify option packets for realtime peers never stopping.
   The option packets not only never stopped, but if a realtime peer was not in
   the peer list multiple options dialogs could accumulate over time.
   (Closes issue #16382. Reported by lftsy. Tested by zerohalo. Patched by
   jpeeler)

* Fix issue where it is possible to crash Asterisk by feeding the curl engine
   invalid data.
   (Closes issue #18161. Reported by wdoekes. Patched by tilghman)

For a full list of changes in this release, please see the ChangeLog:

http://downloads.asterisk.org/pub/telephony/asterisk/ChangeLog-1.8.1